03Develop/BVASPGEVL5001v05ModuloInventarioJdbc/src/pe/bbva/spring5/evalua/modulo/inventario/entity/CE02Product.java
package pe.bbva.spring5.evalua.modulo.inventario.entity;

import java.io.Serializable;

public class CE02Product implements Serializable{
	/**
	 * 
	 */
	private static final long serialVersionUID = 1L;
	private String description;
    private Double price;
    
	public String getDescription() {
		return description;
	}
	public void setDescription(String description) {
		this.description = description;
	}
	public Double getPrice() {
		return price;
	}
	public void setPrice(Double price) {
		this.price = price;
	}
	@Override
	public String toString() {
		StringBuilder builder = new StringBuilder();
		builder.append("CEProduct [description=");
		builder.append(description);
		builder.append(", price=");
		builder.append(price);
		builder.append("]");
		return builder.toString();
	}
}
